+++ lib/fast_ignore.rb
@@ -3,12 +3,14 @@
 require_relative './fast_ignore/backports'
 
 require 'set'
+require 'strscan'
 require_relative './fast_ignore/rule_sets'
 require_relative './fast_ignore/rule_set'
 require_relative './fast_ignore/global_gitignore'
 require_relative './fast_ignore/rule_builder'
 require_relative './fast_ignore/gitignore_rule_builder'
 require_relative './fast_ignore/rule'
+require_relative './fast_ignore/unmatchable_rule'
 require_relative './fast_ignore/shebang_rule'
 
 class FastIgnore